+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles/main.scss';
-import renderAddScore from './modules/domDisplayAddScore.js';
-import renderScore from './modules/domDisplayScores.js';
+import renderAddScore from './modules/displayAddScore.js';
+import renderScore from './modules/fetchScores.js';
 
 window.addEventListener('load', () => {
   renderScore();
